mmanon: fix undefined behavior
closes https://github.com/rsyslog/rsyslog/issues/2243

+++ plugins/mmanon/mmanon.c
@@ -1285,7 +1285,7 @@ anonEmbedded(wrkrInstanceData_t *pWrkrData, uchar **msg, int *pLenMsg, int *idx,
 		*hasChanged = 1;
 
 		if(caddresslen != iplen) {
-			*pLenMsg = *pLenMsg + (caddresslen - iplen);
+			*pLenMsg = *pLenMsg + ((int)caddresslen - (int)iplen);
 			*msg = (uchar*) malloc(*pLenMsg);
 			memcpy(*msg, msgcpy, *idx);
 		}
